@@ -6,33 +6,43 @@
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
     "test:unit": "vue-cli-service test:unit",
-    "lint": "vue-cli-service lint"
+    "lint": "vue-cli-service lint",
+    "vite": "vite",
+    "vite:build": "vite build"
   },
   "dependencies": {
     "core-js": "^3.6.5",
-    "vue": "^2.6.11",
-    "vue-router": "^3.2.0",
-    "vuex": "^3.4.0"
+    "jest": "^26.4.2",
+    "ky": "^0.23.0",
+    "lodash": "^4.17.20",
+    "pinia": "^2.0.0-alpha.2",
+    "vite": "1.0.0-rc.4",
+    "vue": "^3.0.0",
+    "vue-i18n": "9.0.0-beta.3",
+    "vue-router": "4.0.0-beta.11"
   },
   "devDependencies": {
+    "@testing-library/jest-dom": "^5.11.4",
     "@types/jest": "^24.0.19",
-    "@typescript-eslint/eslint-plugin": "^2.33.0",
-    "@typescript-eslint/parser": "^2.33.0",
+    "@types/lodash": "^4.14.161",
+    "@typescript-eslint/eslint-plugin": "^4.2.0",
+    "@typescript-eslint/parser": "^4.2.0",
     "@vue/cli-plugin-babel": "^4.5.0",
     "@vue/cli-plugin-eslint": "^4.5.0",
     "@vue/cli-plugin-typescript": "^4.5.0",
     "@vue/cli-plugin-unit-jest": "^4.5.0",
     "@vue/cli-service": "^4.5.0",
+    "@vue/compiler-sfc": "^3.0.0",
     "@vue/eslint-config-prettier": "^6.0.0",
     "@vue/eslint-config-typescript": "^5.0.2",
-    "@vue/test-utils": "^1.0.3",
-    "eslint": "^6.7.2",
+    "@vue/test-utils": "2.0.0-beta.5",
+    "eslint": "^7.9.0",
     "eslint-plugin-prettier": "^3.1.3",
     "eslint-plugin-vue": "^6.2.2",
-    "prettier": "^1.19.1",
+    "prettier": "^2.1.2",
     "sass": "^1.26.5",
-    "sass-loader": "^8.0.2",
-    "typescript": "~3.9.3",
-    "vue-template-compiler": "^2.6.11"
+    "sass-loader": "^10.0.2",
+    "typescript": "4.0.3",
+    "vue-jest": "^5.0.0-alpha.4"
   }
 }
