@@ -1,3 +1,7 @@
 module.exports = {
-  preset: "@vue/cli-plugin-unit-jest/presets/typescript-and-babel"
+  preset: "@vue/cli-plugin-unit-jest/presets/typescript-and-babel",
+  setupFilesAfterEnv: ["./tests/unit/jest.setup.ts"],
+  transform: {
+    "^.+\\.vue$": "vue-jest"
+  }
 };
